Cleanup and add executor's token balances. Add support for timeouts
