+++ scripts/grabConvexAPY.ts
@@ -1,10 +1,8 @@
 import { BigNumber } from "ethers";
 import { ethers, network } from "hardhat"
-import https from "https";
-import fs from "fs";
 import { ICurveCVXETH, ICurvePoolUSD, ICvxBaseRewardPool } from "../typechain";
 import { reset } from "@nomicfoundation/hardhat-network-helpers";
-
+import { getTokenPrice } from "./common";
 type rewards = {
     tokenAddress: string,
     rewardPerToken: BigNumber
@@ -188,45 +186,3 @@ async function resetBlockBackDays(days: number): (Promise<void>) {
     await reset(process.env.NODE_URL, blockNumberRequested)
 }
 
-async function getTokenPrice(tokenAddress: string, network: string): (Promise<number>) {
-    if (tokenAddress == "0xEeeeeEeeeEeEeeEeEeEeeEEEeeeeEeeeeeeeEEeE") {
-        tokenAddress = "0xc02aaa39b223fe8d0a0e5c4f27ead9083c756cc2"
-    }
-    let url = `https://api.coingecko.com/api/v3/coins/${network}/contract/${tokenAddress}`;
-
-    // Coingecko keeps rate limiting randomly.
-    await delay(5000);
-
-    const getPrice = () => {
-        return new Promise((resolve, reject) => {
-            https.get(url, (resp) => {
-                let data = "";
-                resp.on("data", (chunk) => {
-                    data += chunk;
-                });
-                resp.on("end", () => {
-                    const price = JSON.parse(data);
-                    resolve(price.market_data.current_price.usd);
-                });
-                resp.on("error", (err) => {
-                    reject(err)
-                })
-            });
-        })
-    }
-    let price = undefined;
-    while (price == undefined) {
-        try {
-            price = await getPrice() as number;
-        } catch (err) {
-            console.log("Errored out, retrying");
-            await delay(5000)
-        }
-    }
-
-    return price;
-}
-
-const delay = (delayInms: number) => {
-    return new Promise(resolve => setTimeout(resolve, delayInms));
-}
